backend/requirements.txt
fastapi
uvicorn
python-multipart
datapizza-ai
datapizza-ai-clients-openai-like
datapizza-ai-clients-google
llama-cpp-python
diffusers
torch
transformers
pillow
numpy
huggingface_hub
accelerate